# api-gateway/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Shopping Application API
  version: 1.0.0
  description: Gateway-facing API contract for product, order, and inventory services.
servers:
  - url: http://localhost:9003
security:
  - basicAuth: []
paths:
  /api/product:
    post:
      summary: Create a product
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
    get:
      summary: List products
      tags: [Products]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: name
            enum: [name, skuCode, price]
        - name: query
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Products fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedProductApiResponse"
  /api/product/{id}:
    get:
      summary: Get product by id
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product fetched
          content:
            application/json:
                schema:
                  $ref: "#/components/schemas/ProductApiResponse"
    put:
      summary: Update a product
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
    delete:
      summary: Delete a product
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Product deleted
  /api/product/sku/{skuCode}:
    get:
      summary: Get product by SKU
      tags: [Products]
      parameters:
        - name: skuCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
  /api/order:
    post:
      summary: Place an order
      tags: [Orders]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 100
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
                schema:
                  $ref: "#/components/schemas/OrderApiResponse"
    get:
      summary: List orders
      tags: [Orders]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Orders fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListApiResponse"
  /api/order/{orderNumber}:
    get:
      summary: Get an order
      tags: [Orders]
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderApiResponse"
  /api/order/{orderNumber}/cancel:
    post:
      summary: Cancel a confirmed order
      tags: [Orders]
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderApiResponse"
  /api/inventory:
    get:
      summary: Check inventory by SKU
      tags: [Inventory]
      parameters:
        - name: skuCode
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        "200":
          description: Inventory fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryApiResponse"
  /api/notification:
    get:
      summary: List stored notifications
      tags: [Notifications]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Notifications fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationApiResponse"
  /api/notification/{id}/read:
    patch:
      summary: Mark a notification as read
      tags: [Notifications]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses: { "200": { description: Notification marked read } }
  /api/notification/read-all:
    patch:
      summary: Mark all owned notifications as read
      tags: [Notifications]
      responses: { "200": { description: Notifications marked read } }
  /api/notification/preferences:
    get:
      summary: Get notification preferences
      tags: [Notifications]
      responses: { "200": { description: Preferences fetched } }
    put:
      summary: Update notification preferences
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderConfirmed, orderCancelled, orderFailed]
              properties:
                orderConfirmed: { type: boolean }
                orderCancelled: { type: boolean }
                orderFailed: { type: boolean }
      responses: { "200": { description: Preferences updated } }
  /api/cart:
    get:
      summary: Get the authenticated user's cart
      tags: [Cart]
      responses: { "200": { description: Cart fetched } }
    delete:
      summary: Clear the authenticated user's cart
      tags: [Cart]
      responses: { "200": { description: Cart cleared } }
  /api/cart/items/{skuCode}:
    put:
      summary: Add or replace a cart item
      tags: [Cart]
      parameters:
        - { name: skuCode, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties: { quantity: { type: integer, minimum: 1, maximum: 99 } }
      responses: { "200": { description: Cart updated } }
    delete:
      summary: Remove a cart item
      tags: [Cart]
      parameters:
        - { name: skuCode, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: Cart updated } }
  /api/checkout:
    post:
      summary: Checkout the authenticated user's cart
      tags: [Checkout]
      parameters:
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, maxLength: 100 } }
      responses: { "200": { description: Checkout completed } }
  /actuator/health:
    get:
      summary: Service health
      tags: [Health]
      responses:
        "200":
          description: Health status
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
    ProductRequest:
      type: object
      required: [skuCode, name, price]
      properties:
        skuCode:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: decimal
          minimum: 0.01
    ProductResponse:
      type: object
      properties:
        id:
          type: string
        skuCode:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: decimal
    ProductApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ProductResponse"
    PagedProductResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProductResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
    PagedProductApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/PagedProductResponse"
    OrderRequest:
      type: object
      required: [orderLineDtoList]
      properties:
        orderLineDtoList:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: "#/components/schemas/OrderLineDto"
    OrderLineDto:
      type: object
      required: [skuCode, quantity]
      properties:
        skuCode:
          type: string
        price:
          type: number
          format: decimal
          minimum: 0.01
          description: Ignored when supplied; the product service provides the authoritative price.
        quantity:
          type: integer
          minimum: 1
    OrderResponse:
      type: object
      properties:
        orderNumber:
          type: string
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, FAILED, COMPENSATION_REQUIRED, CANCELLATION_PENDING]
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineDto"
        message:
          type: string
    OrderApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/OrderResponse"
    OrderListApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/OrderResponse"
    InventoryResponse:
      type: object
      properties:
        skuCode:
          type: string
        quantity:
          type: integer
        inStock:
          type: boolean
    InventoryApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/InventoryResponse"
    NotificationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        eventId:
          type: string
        orderNumber:
          type: string
        eventType:
          type: string
        message:
          type: string
        receivedAt:
          type: string
          format: date-time
    PagedNotificationResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/NotificationResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    NotificationApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/PagedNotificationResponse"